% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df}
\alias{df}
\title{Title}
\usage{
df(df, i, Indice)
}
\arguments{
\item{df}{df}

\item{i}{Paper_ID 1-18}

\item{Indice}{indice ("rc", "acc", "dp", "eff", "ddmv", "ddmz")}
}
\value{
结果
}
\description{
Title
}
