% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Title}
\usage{
icc(df, indice, Target)
}
\arguments{
\item{df}{df}

\item{indice}{indice ("rc", "acc", "dp", "eff", "ddmv", "ddmz")}

\item{Target}{target}
}
\value{
结果
}
\description{
Title
}
