% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_dp.R
\name{icc_dp}
\alias{icc_dp}
\title{Title}
\usage{
icc_dp(
  df,
  Target,
  Subject = "Subject",
  Match = "Match",
  Identity = "Identity",
  Session = "Session",
  RT_ms = "RT_ms",
  ACC = "ACC",
  Self = "Self"
)
}
\arguments{
\item{df}{df}

\item{Target}{target}

\item{Subject}{subject id}

\item{Match}{match}

\item{Identity}{identity}

\item{Session}{session}

\item{RT_ms}{rt}

\item{ACC}{acc}

\item{Self}{self}
}
\value{
output
}
\description{
Title
}
