% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_shr_dp.R
\name{loop_shr_dp}
\alias{loop_shr_dp}
\title{Title}
\usage{
loop_shr_dp(list, Target, Paper_ID, Indice, nc)
}
\arguments{
\item{list}{list of mc shr}

\item{Target}{Friend Stranger and ...}

\item{Paper_ID}{Paper_ID}

\item{Indice}{indice ("rc", "acc", "dp", "eff", "ezddm", "rwddm")}

\item{nc}{number of cores}
}
\value{
结果
}
\description{
Title
}
