% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcshr.R
\name{mcshr}
\alias{mcshr}
\title{Title}
\usage{
mcshr(list, nc, indice, Target)
}
\arguments{
\item{list}{list}

\item{nc}{number of core}

\item{indice}{indice ("rc", "acc", "dp", "eff", "ddmv", "ddmz")}

\item{Target}{target}
}
\value{
结果
}
\description{
Title
}
