% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcshr_eff.R
\name{mcshr_eff}
\alias{mcshr_eff}
\title{Title}
\usage{
mcshr_eff(
  list,
  nc,
  Target,
  Subject = "Subject",
  Match = "Match",
  Identity = "Identity",
  Session = "Session",
  RT_ms = "RT_ms",
  ACC = "ACC",
  Self = "Self"
)
}
\arguments{
\item{list}{list}

\item{nc}{number of core}

\item{Target}{target}

\item{Subject}{subject id}

\item{Match}{match}

\item{Identity}{identity}

\item{Session}{session}

\item{RT_ms}{rt}

\item{ACC}{acc}

\item{Self}{self}
}
\value{
结果
}
\description{
Title
}
