% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcshr_rt.R
\name{mcshr_rt}
\alias{mcshr_rt}
\title{Title}
\usage{
mcshr_rt(list, Target, Paper_ID, nc)
}
\arguments{
\item{list}{list of mc shr}

\item{Target}{Friend Stranger and ...}

\item{Paper_ID}{Paper_ID}

\item{nc}{number of cores}
}
\value{
结果
}
\description{
Title
}
