% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcshr_rwddm.R
\name{mcshr_rwddm}
\alias{mcshr_rwddm}
\title{Title}
\usage{
mcshr_rwddm(list, Target, Paper_ID)
}
\arguments{
\item{list}{list}

\item{Target}{target}

\item{Paper_ID}{Paper_ID}
}
\value{
结果
}
\description{
Title
}
