% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_mcshr_rwddm.R
\name{mt_mcshr_rwddm}
\alias{mt_mcshr_rwddm}
\title{Title}
\usage{
mt_mcshr_rwddm(list, Target, Paper_ID, nc)
}
\arguments{
\item{list}{list of mc shr}

\item{Target}{Friend Stranger and ...}

\item{Paper_ID}{Paper_ID}

\item{nc}{number of cores}
}
\value{
结果
}
\description{
Title
}
