% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmshr.R
\name{nmshr}
\alias{nmshr}
\title{Title}
\usage{
nmshr(list, indice, Target)
}
\arguments{
\item{list}{list}

\item{indice}{indice ("rc", "acc", "dp", "eff", "ddmv", "ddmz")}

\item{Target}{target}
}
\value{
结果
}
\description{
Title
}
