% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmshr_ddmz.R
\name{nmshr_ddmz}
\alias{nmshr_ddmz}
\title{Title}
\usage{
nmshr_ddmz(list, Target, Paper_ID)
}
\arguments{
\item{list}{list}

\item{Target}{target}

\item{Paper_ID}{Paper_ID}
}
\value{
结果
}
\description{
Title
}
