% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmshr_eff.R
\name{nmshr_eff}
\alias{nmshr_eff}
\title{Title}
\usage{
nmshr_eff(list, Target, Paper_ID)
}
\arguments{
\item{list}{list}

\item{Target}{target}

\item{Paper_ID}{Paper_ID}
}
\value{
结果
}
\description{
Title
}
