% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr.R
\name{shr}
\alias{shr}
\title{Title}
\usage{
shr(list, indice, Target, mc = FALSE, nc = NULL)
}
\arguments{
\item{list}{list}

\item{indice}{indice ("rc", "acc", "dp", "eff", "ddmv", "ddmz")}

\item{Target}{target}

\item{mc}{Monte Carlo or not}

\item{nc}{number of core}
}
\value{
结果
}
\description{
Title
}
