% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcshr.R
\name{mcshr}
\alias{mcshr}
\title{Title}
\usage{
mcshr(list, indice, Target, Paper_ID)
}
\arguments{
\item{list}{list}

\item{indice}{indice ("rc", "acc", "dp", "eff", "ddmv", "ddmz")}

\item{Target}{target}

\item{Paper_ID}{Paper_ID}
}
\value{
结果
}
\description{
Title
}
