% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcshr.R
\name{mcshr}
\alias{mcshr}
\title{Title}
\usage{
mcshr(list, Target, Paper_ID, Indice, nc)
}
\arguments{
\item{list}{list}

\item{Target}{target}

\item{Paper_ID}{Paper_ID}

\item{Indice}{indice ("rc", "acc", "dp", "eff", "ezddm", "rwddm")}

\item{nc}{number of cores}
}
\value{
结果
}
\description{
Title
}
