% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr.R
\name{shr}
\alias{shr}
\title{Title}
\usage{
shr(list, indice, Target, Paper_ID, mc = FALSE)
}
\arguments{
\item{list}{list}

\item{indice}{indice ("rc", "acc", "dp", "eff", "ddmv", "ddmz")}

\item{Target}{target}

\item{Paper_ID}{Paper_ID}

\item{mc}{Monte Carlo or not}
}
\value{
结果
}
\description{
Title
}
