% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr.R
\name{shr}
\alias{shr}
\title{Title}
\usage{
shr(list, Target, Paper_ID, Indice, mc = FALSE, nc = 3)
}
\arguments{
\item{list}{对象}

\item{Target}{Identity有几种情况}

\item{Paper_ID}{文章的序号}

\item{Indice}{这次分析的指标是什么indice ("rc", "acc", "dp", "eff", "ezddm", "rwddm")}

\item{mc}{是否是蒙特卡洛的list}

\item{nc}{需要用多少个核心。普通分半建议3个，蒙特卡洛分半则使用最大值}
}
\value{
结果
}
\description{
Title
}
